remove es6 and typescript version of setup:karma command
