--- bin/init/scripts/karma/es6.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-'use strict';
